requirements.txt
flask=3.1.0
pandas=2.2.3
plotly=6.0.0
python-dotenv=1.0.1
kaleido=0.2.1